Initial routing done
Did the routing for the Sidebar links.
